refactor: outsourced the sortProjects function in a single util
